fix:galeryOk
