+++ museum/assets/jscript/index.js
@@ -1,0 +1,5 @@
+import initComparisons from './comparisons.js';
+import changeSlide from './wgallery.js';
+
+
+initComparisons();
