+++ museum/index.html
@@ -265,6 +265,22 @@
        
         <div class="picture-container">
             <div class="picture-inner-container">
+                <img src="./assets/img/galery/galery11.jpg" alt="galery11">
+                <img src="./assets/img/galery/galery12.jpg" alt="galery12">
+                <img src="./assets/img/galery/galery8.jpg" alt="galery8">
+                <img src="./assets/img/galery/galery9.jpg" alt="galery9">
+                <img src="./assets/img/galery/galery10.jpg" alt="galery10">
+                <img src="./assets/img/galery/galery1.jpg" alt="galery1">
+                <img src="./assets/img/galery/galery2.jpg" alt="galery2">
+                <img src="./assets/img/galery/galery3.jpg" alt="galery3">
+                <img src="./assets/img/galery/galery4.jpg" alt="galery4">
+                <img src="./assets/img/galery/galery5.jpg" alt="galery6">
+                <img src="./assets/img/galery/galery6.jpg" alt="galery6">
+                <img src="./assets/img/galery/galery7.jpg" alt="galery7">
+                <img src="./assets/img/galery/galery13.jpg" alt="galery13">
+                <img src="./assets/img/galery/galery14.jpg" alt="galery14">
+                <img src="./assets/img/galery/galery15.jpg" alt="galery15">
+
             </div>
           </div>
     </div>   
@@ -385,10 +401,13 @@
         </div>
         </div>
    </footer>
-   <script src="./assets/jscript/main.js">
+   <script type="module" src="./assets/jscript/index.js"></script>
+  <!-- <script src="./assets/jscript/main.js">
    </script>
    <script>
       initComparisons();
    </script>
+-->
+
 </body>
 </html>
